% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_event}
\alias{get_event}
\title{get search options for event}
\usage{
get_event()
}
\value{
character vector of events
}
\description{
return options of event params of \code{\link{search_speech}}.
}
