% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_field}
\alias{get_field}
\title{get search options for field}
\usage{
get_field()
}
\value{
character vector of fields
}
\description{
return options of field params of \code{\link{search_speech}}.
}
