% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_president}
\alias{get_president}
\title{get search options for president}
\usage{
get_president()
}
\value{
character vector of president names
}
\description{
return options of president params of \code{\link{search_speech}}.
}
