% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speech.R
\name{get_speech}
\alias{get_speech}
\title{get speech}
\usage{
get_speech(link, paragraph = FALSE)
}
\arguments{
\item{link}{Sciprt link form \code{\link{search_speech}}. Like form is like
\url{http://www.pa.go.kr/research/contents/speech/index.jsp?spMode=view&catid=c_pa02062&artid=1310475}}

\item{paragraph}{logical if sepeate paragraph or not. Default is FALSE.}
}
\value{
a [tibble][tibble::tibble-package] title, date, president, place, field, event, source, paragraph, content
}
\description{
get Speech's script of link from \code{\link{search_speech}}. only can get one script at once.
}
