% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_speech.R
\name{search_speech}
\alias{search_speech}
\title{search speech}
\usage{
search_speech(
  keyword = "",
  searchField = 0,
  count = 10000,
  president = "",
  mediaType = c("doc", "mp3", "mp4", "all"),
  event = "",
  order = 1
)
}
\arguments{
\item{keyword}{keyword to search. Default is all results. Accept character data type.}

\item{searchField}{All:0, Title:1, Content:2, Place:3. Default is 0 means all parts. Accept numeric data type.}

\item{count}{request number of results. Default is all. Accept numeric data type.}

\item{president}{president to search. Default is all results.
\code{\link{get_president}} provides available list. Accept character data type.}

\item{mediaType}{mediaType}

\item{event}{event to search. Default is all results.
\code{\link{get_event}} provides available list. Accept character data type.}

\item{order}{order of search results. Defaults is order by date.
order by date:1, order by title character:2. Accept numeric data type.}
}
\value{
a [tibble][tibble::tibble-package] president, field, event, title, date, link
}
\description{
search script of president's speech. Data is form \href{http://www.pa.go.kr/research/contents/speech/index.jsp}{presidential archives of korea}.
}
\examples{
\dontrun{
splist <- search_speech(count = 20)
splist
}

}
