% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_list.R
\docType{data}
\name{speech_list}
\alias{speech_list}
\title{speech list data}
\format{
tibble with columns
\describe{
\item{president}{president}
\item{field}{field}
\item{event}{event}
\item{title}{title}
\item{date}{date}
\item{link}{link}
}
}
\source{
http://www.pa.go.kr/research/contents/speech/index.jsp
}
\usage{
speech_list
}
\description{
speech list data
}
\examples{
  speech_list
}
\keyword{datasets}
