% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spidx.R
\docType{data}
\name{spidx}
\alias{spidx}
\title{speech index data}
\format{tibble with columns
\describe{
\item{president}{president}
\item{field}{field}
\item{event}{event}
\item{title}{title}
\item{date}{date}
\item{link}{link}
}}
\source{
http://www.pa.go.kr/research/contents/speech/index.jsp
}
\usage{
spidx
}
\description{
speech index data
}
\examples{
  spidx
}
\keyword{datasets}
