% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_options.R
\name{get_event}
\alias{get_event}
\title{#' get search options for field
#'
#' return options of field params of \code{\link{search_speech}}.
#'
#' @return character vector of fields
#'
#' @export
#' @importFrom rvest html_nodes
#' @importFrom rvest html_text
#' @importFrom rvest html_attr
#' @importFrom xml2 read_html}
\usage{
get_event()
}
\value{
character vector of events
}
\description{

}
\details{

}
